Changes the datatype for price from Decimal to Double

+++ Sources/GroceryAppSharedDTO/GroceryItemDTO.swift
@@ -9,10 +9,10 @@ import Foundation
 
 public struct GroceryItemRequestDTO: Codable {
     public var title: String
-    public var price: Decimal
+    public var price: Double
     public var quantity: Int
     
-    public init(title: String, price: Decimal, quantity: Int) {
+    public init(title: String, price: Double, quantity: Int) {
         self.title = title
         self.price = price
         self.quantity = quantity
@@ -22,10 +22,10 @@ public struct GroceryItemRequestDTO: Codable {
 public struct GroceryItemResponseDTO: Codable {
     public var id: UUID
     public var title: String
-    public var price: Decimal
+    public var price: Double
     public var quantity: Int
 
-    public init(id: UUID, title: String, price: Decimal, quantity: Int) {
+    public init(id: UUID, title: String, price: Double, quantity: Int) {
         self.id = id
         self.title = title
         self.price = price
